+++ fast-toolkit/src/test/java/encrypit/RSATest.java
@@ -8,7 +8,9 @@ import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.security.SignatureException;
 
+import com.thankjava.toolkit.bean.encrypit.RSAEncryptAlgorithm;
 import com.thankjava.toolkit.bean.encrypit.RSAKey;
+import com.thankjava.toolkit.bean.encrypit.RSASignAlgorithm;
 import com.thankjava.toolkit.core.encrypit.RSA;
 
 public class RSATest {
@@ -23,10 +25,10 @@ public class RSATest {
         // 私钥加签数据
         String text = "RSA非对称算法";
 
-        byte[] cipherByte = RSA.encrypt(text.getBytes(), pubK);
-        System.out.println("RSA解密：" + new String(RSA.decrypt(cipherByte, priK)));
+//        byte[] cipherByte = RSA.encrypt(text.getBytes(), pubK, RSAEncryptAlgorithm.RSA_ECB_OAEPWithSHA_256AndMGF1Padding);
+//        System.out.println("RSA解密：" + new String(RSA.decrypt(cipherByte, priK, RSAEncryptAlgorithm.RSA_ECB_OAEPWithSHA_256AndMGF1Padding)));
 
-        byte[] signArray = RSA.sign(text.getBytes(), priK);
-        System.out.println("验签结果：" + RSA.verify(text.getBytes(), signArray, pubK));
+        byte[] signArray = RSA.sign(text.getBytes(), priK, RSASignAlgorithm.SHA1withRSA);
+        System.out.println("验签结果：" + RSA.verify(text.getBytes(), signArray, pubK, RSASignAlgorithm.SHA1withRSA));
     }
 }