+++ doc/ChangeLog.md
@@ -6,6 +6,21 @@
     
         - 调整RedisManager初始化可通过redis.properties的redis.db.index属性来指定redis默认数据库
         - JSON模块新增 toFormatJSONString() 输出美化后的json字符串
+        - RSA模块拓展API功能提升
+        
+            ```
+            1. 允许指定加解密 paddming algorithm
+            
+                RSA/ECB/PKCS1Padding
+                RSA/ECB/OAEPWithSHA-1AndMGF1Padding
+                RSA/ECB/OAEPWithSHA-256AndMGF1Padding
+                
+            2. 允许指定加验签 模式 algorithm
+            
+                SHA1withRSA
+                SHA256withRSA
+                
+            ```
 
     - 1.1.2